Normalizer for Skylight integration.
Skylight (skylight.io) is a rails application monitoring service, which
can easily instrument code that integrates with
`ActiveSupport::Notifications`.

For now we will just subscribe to the `request.orientdb_client` event,
which is of greatest interest, as it most closely represents OrientDB
query time.

Ideally we would parse/lex the URL as Skylight does for sql queries, so
that greater information would be available in the `description`, and
also so we could ignore things like property creation and alteration,
but I think this would be a) very difficult to get right, without a
large sample of live OrientDB queries and/or a deep knowledge of its
query language, which I lack; and b) probably fairly resource intensive,
since we'd likely have to do a bunch of regex'ing, which I think would
add undue overhead (bad cost/benefit ratio).